Revert "Merge pull request #44 from OxIonics/tph/defaults"
This reverts commit dde627d467ad9a15ff164f5912557f049cf6b944, reversing
changes made to 77cc00358115e670fb111d2ec40822f9d8429df1.

We have decided to revert to prevous functionality of ndscan, removing the
"Default" scan option for now. This is because the parameter values used
for a given experimental run where no longer visible when the "Default" option
was selected.